view built-in privacy policy from telemetry dialog
